Reduce examples CI compile surface
